Fix upload_textures layer indices
