+++ wasm/src/communication/mod.rs
@@ -1,6 +1,5 @@
 pub mod message_queue;
 pub mod sprite;
-pub mod state;
 pub mod synchronization_memory;
 
 #[doc(inline)]
@@ -11,11 +10,9 @@ use rask_engine::resources;
 #[doc(inline)]
 pub use sprite::Sprite;
 #[doc(inline)]
-use state::State;
-#[doc(inline)]
 pub use synchronization_memory::{GameState, SynchronizationMemory};
 
-pub static DOUBLE_BUFFER: Mutex<State> = Mutex::new(State::empty());
+pub static DOUBLE_BUFFER: Mutex<Vec<Sprite>> = Mutex::new(Vec::new());
 pub static RESOURCE_TABLE: RwLock<resources::ResourceTable> =
     RwLock::new(resources::ResourceTable::new());
 pub static TEXTURE_IDS: Mutex<resources::TextureIds> = Mutex::new(resources::TextureIds::empty());
